Extend models from base abstract class
