+++ src/Models/Prefix.php
@@ -2,36 +2,15 @@
 
 namespace Theutz\Unite\Models;
 
-use Illuminate\Support\Collection;
 use Theutz\Unite\Category;
-use Theutz\Unite\Loader;
 
 /**
- * @mixin Collection
+ * {@inheritDoc}
  */
-class Prefix
+class Prefix extends Model
 {
-    private Collection $data;
-
-    public function __construct(
-        private Loader $loader
-    ) {
-    }
-
-    public static function __callStatic($name, $args)
+    protected function category(): Category
     {
-        $instance = app(self::class);
-
-        return $instance->getData()->$name($args);
-    }
-
-    private function getData(): Collection
-    {
-        if (! isset($this->data)) {
-            $data = $this->loader->load(Category::Unit);
-            $this->data = collect($data);
-        }
-
-        return $this->data;
+        return Category::Prefix;
     }
 }
