+++ src/Models/Unit.php
@@ -7,31 +7,12 @@ use Theutz\Unite\Category;
 use Theutz\Unite\Loader;
 
 /**
- * @mixin Collection
+ * {@inheritDoc}
  */
-class Unit
+class Unit extends Model
 {
-    private Collection $data;
-
-    public function __construct(
-        private Loader $loader
-    ) {
-    }
-
-    public static function __callStatic($name, $args)
+    protected function category(): Category
     {
-        $instance = app(self::class);
-
-        return $instance->getData()->$name($args);
-    }
-
-    private function getData(): Collection
-    {
-        if (! isset($this->data)) {
-            $data = $this->loader->load(Category::Unit);
-            $this->data = collect($data);
-        }
-
-        return $this->data;
+        return Category::Unit;
     }
 }
